refactor(payment): move mediaOffer parsing to usePlaylist and useMedia
