+++ src/components/CardGrid/CardGrid.tsx
@@ -5,7 +5,7 @@ import styles from './CardGrid.module.scss';
 
 import useBreakpoint, { Breakpoint, Breakpoints } from '#src/hooks/useBreakpoint';
 import { chunk, findPlaylistImageForWidth } from '#src/utils/collection';
-import { isLocked } from '#src/utils/cleeng';
+import { isLocked } from '#src/utils/entitlements';
 import Card from '#src/components/Card/Card';
 import VirtualizedGrid from '#src/components/VirtualizedGrid/VirtualizedGrid';
 import type { AccessModel } from '#types/Config';
