+++ src/hooks/usePlaylist.ts
@@ -2,8 +2,9 @@ import { UseBaseQueryResult, useQuery } from 'react-query';
 
 import { generatePlaylistPlaceholder } from '../utils/collection';
 import { getPlaylistById } from '../services/api.service';
+import { filterMediaOffers } from '../utils/entitlements';
 
-import type { Playlist } from '#types/playlist';
+import type { Playlist, PlaylistItem } from '#types/playlist';
 
 const placeholderData = generatePlaylistPlaceholder(30);
 
@@ -16,9 +17,24 @@ export default function usePlaylist(
   usePlaceholderData: boolean = true,
   limit?: number,
 ): UsePlaylistResult {
-  return useQuery(['playlist', playlistId, relatedMediaId], () => getPlaylistById(playlistId, relatedMediaId, limit), {
-    enabled: !!playlistId && enabled,
-    placeholderData: usePlaceholderData ? placeholderData : undefined,
-    retry: false,
-  });
+  return useQuery(
+    ['playlist', playlistId, relatedMediaId],
+    async () => {
+      const playlist = await getPlaylistById(playlistId, relatedMediaId, limit);
+
+      // Parse TVOD media offers, if present
+      if (playlist?.playlist)
+        playlist.playlist = playlist.playlist.map((item: PlaylistItem) => ({
+          ...item,
+          mediaOffers: item.productIds ? filterMediaOffers('cleeng', item.productIds) : undefined,
+        }));
+
+      return playlist;
+    },
+    {
+      enabled: !!playlistId && enabled,
+      placeholderData: usePlaceholderData ? placeholderData : undefined,
+      retry: false,
+    },
+  );
 }
