+++ types/playlist.d.ts
@@ -38,8 +38,9 @@ export type PlaylistItem = {
   title: string;
   tracks: Track[];
   variations?: Record<string, unknown>;
-  productIds?: string;
   free?: string;
+  productIds?: string;
+  mediaOffers?: MediaOffer[] | null;
 };
 
 export type Link = {
